requirements.txt
coverage==7.10.0
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
flake8==7.3.0
inflection==0.5.1
mccabe==0.7.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
redis==6.2.0

